+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, e3nn Developers'
 author = 'e3nn Developers'
 
 # The full version, including alpha/beta/rc tags
-release = '0.3.2'
+release = '0.3.3'
 
 # -- General configuration ---------------------------------------------------